+++ src/App.tsx
@@ -6,6 +6,7 @@ import Home from './Components/home';
 import Footer from './Components/Footer/Footer'
 import Header from './Components/Header/Header'
 import Page1 from './Components/Pages/Page1'
+import Page2 from './Components/Pages/Page2'
 import NotFound from './Components/Errors/NotFound'
 window.onbeforeunload = () => '';
 
@@ -51,6 +52,7 @@ export default class App extends React.Component<DJGUIAppProps, DJGUIAppState> {
               <Route path='/login'>{this.state.jwtToken !== '' ? <Redirect to='/home'/> : <Login setJWTTokenAndHostName={this.setJWTTokenAndHostName}></Login>}</Route>
               <Route path='/home'>{this.state.jwtToken !== '' ? <Home jwtToken={this.state.jwtToken}></Home> : <Redirect to='/login'/>}</Route>
               <Route path='/page1'>{this.state.jwtToken !== '' ? <Page1 jwtToken={this.state.jwtToken}></Page1> : <Redirect to='/login'/>}</Route>
+              <Route path='/page2'>{this.state.jwtToken !== '' ? <Page2 jwtToken={this.state.jwtToken}></Page2> : <Redirect to='/login'/>}</Route>
               <Route path="*" component={NotFound} />
             </Switch>
           </div>
