+++ src/Components/Pages/Page2.tsx
@@ -17,9 +17,12 @@ interface Page1Props {
             <SideBar />
             <ul>
               <li style={{display: 'block'}}> 
-                <GridLayout className="mygrid" cols={ 1 }    rowHeight={400} width={1600}>
+                <GridLayout className="mygrid" cols={ 3 }    rowHeight={400} width={1600}>
                   <div key='component1' data-grid={{x: 0, y: 0, w: 1, h: 1, static: true}}>
-                  <TableView token={this.props.jwtToken} route='/query1' tableName='/query1'/>
+                  <TableView token={this.props.jwtToken} route='/query2' tableName='/query2'/>
+                  </div>
+                  <div key='component2' data-grid={{x: 1, y: 0, w: 1, h: 1, static: true}}>
+                  <TableView token={this.props.jwtToken} route='/query3' tableName='/query3'/>
                   </div>
                 </GridLayout>
               </li>
