+++ test_spec.yaml
@@ -1,0 +1,79 @@
+version: 'v0.0.0'
+LabBook: null
+SciViz: # top level tab
+  auth:
+    mode: database # reuse from DJ LabBook
+  pages: # individual pages
+    Page1: # name reference of the page
+      route: /sessions # dev, be careful of name collisions
+      args: null
+      hidden: false
+      grids:
+        grid1:
+          type: fixed
+          max_component_per_page: 1
+          columns: 1
+          row_height: 400
+          row_width: 1600
+          components:
+            component1:
+              route: /query1
+              x: 0 # component indexes at top left corner
+              y: 0
+              height: 1
+              width: 1
+              type: entry-browser # entry-browser|meta-info|plot:png|plot:plotly1|plot:plotly2|plot:hybrid|plot:table|raw-text
+              restriction: >
+                def restriction(**args):
+                  return dict()
+              dj_query: > # assume all schemas that user has access will be VM loaded
+                def dj_query():
+                  chen_lab.spawn_missing_classes()
+                  return HomeCage * ModuleEvent * EventType, , dict(order_by=['KEY DESC', 'mouse_id'], limit=15)
+              redirect_page: 'page2'
+    Page2:
+      route: /session # dev, be careful of name collisions
+      args:
+        - subject_name
+        - session_number
+      hidden: true
+      grids:
+        grid1:
+          type: fixed
+          max_component_per_page: 20
+          columns: 3
+          row_height: 400
+          row_width: 1600
+          components:
+            component1:
+              route: /query2
+              x: 0
+              y: 0
+              height: 1
+              width: 1
+              type: plot:png
+              restriction: >
+                def restriction(**args):
+                  return dict(**args)
+              dj_query: >
+                def dj_query():
+                  chen_lab.spawn_missing_classes()
+                  return TrainingStats.proj(png='trial_mean_png'), dict(order_by=['KEY DESC', 'mouse_id'], limit=15)
+            component2:
+              route: /query3
+              x: 1
+              y: 0
+              height: 1
+              width: 1
+              type: plot:plotly1
+              restriction: >
+                def restriction(**args):
+                  return dict(**args)
+              dj_query: >
+                def dj_query():
+                  chen_lab.spawn_missing_classes()
+                  return TrainingStatsPlotly.proj(plotly1='trial_mean'), dict(order_by=['KEY DESC', 'mouse_id'], limit=15)  # (dj_query() & restriction()).fetch1('plotly1')
+              diff_checker: >
+                def diff_checker(**args):
+                  return TrainingStatsPlotly.proj(hash='trial_mean_hash')
+
